Refactoring fill database

+++ functions.py
@@ -2,6 +2,7 @@ import re
 import sqlite3
 import tomllib
 from pathlib import Path
+from datetime import datetime
 
 from classes import Rule
 from patterns import UPPER_OBJ, UPPER_L3_PROTO, UPPER_L4_PROTO
@@ -9,23 +10,52 @@ from patterns import SUBRULE_L3, SUBRULE_L4
 
 
 def init_database(config: dict):
-    database = Path(config.get('database'))
-    if database.exists():
-        confirm = input('File db exists, delete? [y|N]: ')
-        if confirm.lower().startswith('y'):
-            database.unlink()
-        else:
-            exit()
-    print('Creating file db...')
-    with sqlite3.connect(config['database']) as conn:
-        with open('sql/sql_execute_before.sql') as f:
-            queryes = f.read()
-        cursor = conn.cursor()
-        cursor.executescript(queryes)
-        conn.commit()
+    '''
+    Функция инициализации БД
+    '''
+    if config['create_sql_dump']:
+        queryset = Path(config.get('queryset_file'))
+        if queryset.exists():
+            confirm = input('File queryset exists, delete? [y|N]: ')
+            if confirm.lower().startswith('y'):
+                try:
+                    queryset.unlink()
+                except FileNotFoundError as e:
+                    print(e)
+                with open(config['queryset_file'], 'w') as f:
+                    f.write(f'-- DATASET: {datetime.today()} --\n')
+            else:
+                config['queryset_memory'] = []
+    else:
+        # Создаем список для хранения селектов
+        config['queryset_memory'] = []
+    if config['init_db']:
+        database = Path(config.get('database'))
+        if database.exists():
+            confirm = input('File db exists, delete? [y|N]: ')
+            if confirm.lower().startswith('y'):
+                try:
+                    database.unlink()
+                except FileNotFoundError as e:
+                    print(e)
+
+                print('Creating file db...')
+                with sqlite3.connect(config['database']) as conn:
+                    with open('sql/sql_execute_before.sql') as f:
+                        queryes = f.read()
+                    cursor = conn.cursor()
+                    cursor.executescript(queryes)
+                    conn.commit()
+            else:
+                print('db not create, use old data')
+                config['init_db'] = False
 
 
 def init_app(config: str = 'asa_duprules.toml') -> dict:
+    '''
+    Функция инициализации приложения
+    читаем конфиг, подготавливаем БД
+    '''
     with open(config, 'rb') as f:
         cfg: dict = tomllib.load(f)
         # load rules to config var
@@ -33,7 +63,9 @@ def init_app(config: str = 'asa_duprules.toml') -> dict:
         with open(sh_access_list, 'r') as f:
             rules = f.read().split('\n')
             cfg['rules'] = rules
-    # init_database(cfg)
+    init_database(cfg)
+    if cfg['init_db']:
+        fill_db(cfg)
     return cfg
 
 
@@ -41,19 +73,16 @@ def define_rule(line: str, config):
     # Верхнеуровневое правило
     if line.startswith('access-list'):
         if re_match := re.fullmatch(UPPER_OBJ, line):
-            ins_obj(re_match, config, 'upper_rules')
-            pass
+            create_query(re_match, config, 'upper_rules')
         if re_match := re.fullmatch(UPPER_L3_PROTO, line):
-            ins_obj(re_match, config, 'upper_rules')
-            pass
+            create_query(re_match, config, 'upper_rules')
         if re_match := re.fullmatch(UPPER_L4_PROTO, line):
-            ins_obj(re_match, config, 'upper_rules')
-            pass
+            create_query(re_match, config, 'upper_rules')
     if line.startswith(' '):
         if re_match := re.fullmatch(SUBRULE_L3, line):
-            ins_obj(re_match, config, 'rules')
+            create_query(re_match, config, 'rules')
         if re_match := re.fullmatch(SUBRULE_L4, line):
-            ins_obj(re_match, config, 'rules')
+            create_query(re_match, config, 'rules')
 
 
 def ins_obj(re_match: re.Match, config: dict, table: str):
@@ -71,13 +100,30 @@ def ins_obj(re_match: re.Match, config: dict, table: str):
         conn.commit()
 
 
+def create_query(re_match: re.Match, config: dict, table: str):
+    finded_objects = {k: v for k, v in re_match.groupdict().items() if v is not None}
+    query = (
+        f'INSERT INTO {table} ('
+        f'{", ".join(finded_objects.keys())}'
+        ') VALUES ("'
+        f'{"\", \"".join(finded_objects.values())}'
+        f'");'
+    )
+    if config['create_sql_dump']:
+        with open(config['queryset_file'], 'a') as f:
+            f.write(query + '\n')
+    else:
+        config['queryset_memory'].append(query)
+
+
 def create_rules_bulk(config: dict):
+    '''
+    Создаем словарь с объектами правил
+    '''
     rules = []
     database = config.get('database')
-    with sqlite3.connect(database) as conn:
-        cursor = conn.cursor()
-        rows = cursor.execute(
-            (
+    # Запрос правил верхнего уровня без групповых объектов
+    query_upper = (
                 'SELECT * FROM upper_rules '
                 'WHERE obj_gr_serv IS NULL AND '
                 'obj_gr_src IS NULL AND '
@@ -85,23 +131,34 @@ def create_rules_bulk(config: dict):
                 'obj_gr_dst IS NULL AND '
                 'object_dst IS NULL;'
             )
-        )
-        fields = [i[0] for i in cursor.description]
-        for row in rows:
-            raw_data = {k: v for k, v in zip(fields, row) if v is not None}
-            rules.append(Rule(raw_data))
-    with sqlite3.connect(database) as conn:
-        cursor = conn.cursor()
-        rows = cursor.execute('SELECT * FROM rules;')
-        fields = [i[0] for i in cursor.description]
-        for row in rows:
-            raw_data = {k: v for k, v in zip(fields, row) if v is not None}
-            rules.append(Rule(raw_data))
+    # Запрос правил нижнего уровня (групповых объектов в них нет)
+    query_lower = 'SELECT * FROM rules;'
+
+    for query in (query_upper, query_lower):
+        with sqlite3.connect(database) as conn:
+            cursor = conn.cursor()
+            rows = cursor.execute(query)
+            fields = [i[0] for i in cursor.description]
+            for row in rows:
+                raw_data = {k: v for k, v in zip(fields, row) if v is not None}
+                rules.append(Rule(raw_data))
     return rules
 
 
-def main(config: dict):
+def fill_db(config: dict):
+    print('Fill database...')
     for line_rule in config['rules']:
-        line = define_rule(line_rule, config)
-        if line:
-            result += 1
+        define_rule(line_rule, config)
+
+    database = config.get('database')
+
+    with sqlite3.connect(database) as conn:
+        cursor = conn.cursor()
+        if config['create_sql_dump']:
+            with open(config['queryset_file']) as f:
+                queryset = f.read()
+            cursor.executescript(queryset)
+        else:
+            query = '\n'.join(config['queryset_memory'])
+            cursor.executescript(query)
+        conn.commit()
